server/api/v1/requirements.txt
django==4.0
djangorestframework
psycopg2-binary
django-environ
django-cors-headers